vibetrading/exchanges/extended.py
"""
X10 Extended exchange sandbox for live trading on StarkNet.

Requires: pip install vibetrading[extended]
  (x10-python-trading)
"""

import logging
from typing import Dict, List, Any, Optional

import pandas as pd

from .base import LiveSandboxBase
from ..models.orders import (
    PerpAccountSummary, PerpPositionSummary,
    CancelOrdersResponse, PerpOrderResponse, PerpOrder,
)
from ..utils.notification import NotificationDeduplicator

logger = logging.getLogger(__name__)

try:
    from x10.perpetual.accounts import StarkPerpetualAccount
    from x10.perpetual.trading_client import PerpetualTradingClient
    from x10.perpetual.stream_client import PerpetualStreamClient
    from x10.perpetual.configuration import STARKNET_MAINNET_CONFIG
    from x10.perpetual.orders import OrderSide
    _HAS_X10 = True
except ImportError:
    _HAS_X10 = False


class ExtendedSandbox(LiveSandboxBase):
    """Live trading sandbox for X10 Extended on StarkNet (perps only)."""

    def __init__(
        self,
        api_key: Optional[str] = None,
        api_secret: Optional[str] = None,
        mode: str = "live",
        address: str = "",
        public_key: str = "",
        private_key: str = "",
        vault: int = 0,
        notification_deduplicator: Optional[NotificationDeduplicator] = None,
        **kwargs,
    ):
        if not _HAS_X10:
            raise ImportError(
                "x10 SDK not installed. Install with: "
                "pip install vibetrading[extended]"
            )

        super().__init__(
            exchange_name="extended",
            api_key=api_key,
            api_secret=api_secret,
            mode=mode,
            notification_deduplicator=notification_deduplicator,
            **kwargs,
        )
        self.account_address = address

        self.stark_account = StarkPerpetualAccount(
            vault=vault, private_key=private_key,
            public_key=public_key, api_key=api_key,
        )
        self.client = PerpetualTradingClient(STARKNET_MAINNET_CONFIG, self.stark_account)
        self.stream_client = PerpetualStreamClient(api_url=STARKNET_MAINNET_CONFIG.stream_url)

        self.orderbook_cache: Dict[str, Dict[str, Any]] = {}
        logger.info("ExtendedSandbox ready (mode=%s)", mode)

    # -- Minimal stub implementations --------------------------------
    # A full implementation would map each method to the X10 API.
    # Below are placeholders that match the VibeSandboxBase interface.

    def get_price(self, asset: str) -> float:
        raise NotImplementedError("get_price not yet implemented for Extended")

    def get_spot_price(self, asset: str) -> float:
        raise NotImplementedError("Spot not supported on Extended")

    def get_perp_price(self, asset: str) -> float:
        return self.get_price(asset)

    def my_spot_balance(self, asset: str) -> float:
        return 0.0

    def my_futures_balance(self, asset: str = "USDC") -> float:
        raise NotImplementedError

    def buy(self, asset, quantity, price, order_type="limit"):
        raise NotImplementedError("Spot trading not supported on Extended")

    def sell(self, asset, quantity, price, order_type="limit"):
        raise NotImplementedError("Spot trading not supported on Extended")

    def long(self, asset, quantity, price, order_type="limit"):
        raise NotImplementedError

    def short(self, asset, quantity, price, order_type="limit"):
        raise NotImplementedError

    def reduce_position(self, asset, quantity):
        raise NotImplementedError

    def set_leverage(self, asset, leverage):
        raise NotImplementedError

    def get_futures_position(self, asset):
        raise NotImplementedError

    def get_spot_ohlcv(self, asset, interval, limit):
        return pd.DataFrame()

    def get_futures_ohlcv(self, asset, interval, limit):
        raise NotImplementedError

    def get_funding_rate(self, asset="BTC", timestamp=None):
        return 0.0

    def get_funding_rate_history(self, asset, limit):
        return pd.DataFrame(columns=["timestamp", "fundingRate"])

    def get_open_interest(self, asset="BTC", timestamp=None):
        return 0.0

    def get_open_interest_history(self, asset, limit):
        return pd.DataFrame(columns=["timestamp", "openInterest"])

    def cancel_order(self, order_id):
        raise NotImplementedError

    def get_perp_open_orders(self, asset=None):
        return []

    def get_spot_open_orders(self, asset=None):
        return []

    def cancel_spot_orders(self, asset, order_ids):
        return CancelOrdersResponse(status="success", orders=[]).to_dict()

    def cancel_perp_orders(self, asset, order_ids):
        raise NotImplementedError

    def get_perp_summary(self):
        return PerpAccountSummary().to_dict()

    def get_perp_position(self, asset):
        return None

    def get_spot_summary(self):
        from ..models.orders import SpotAccountSummary
        return SpotAccountSummary().to_dict()

    def get_futures_unrealized_pnl(self, asset=None):
        return 0.0

    def get_total_futures_unrealized_pnl(self):
        return 0.0

    def get_available_margin(self):
        return 0.0

    def get_total_margin_used(self):
        return 0.0

    def apply_funding_payment(self, asset, position_size, leverage=1):
        pass

    def advance_time(self, interval):
        return None
